Added Firebase Contexts that allow for less props and global variables.
There is a new firebase class that holds the database and the auth for the Firebase. Using props.firebase.db will get the firebase database in the Consumer context.
