+++ src/Note.js
@@ -1,11 +1,14 @@
 import React from 'react';
 import './App.css';
 import ProgressStepper from './Stepper'
+import { FireBaseContext } from './assets/components/FireBase';
 
 function Note() {
   return (
     <div>
-      <ProgressStepper className="stepper"/>
+      <FireBaseContext.Consumer>
+        {firebase => <ProgressStepper firebase={firebase} className="stepper"/>}
+      </FireBaseContext.Consumer>
     </div>
   );
 }
