+++ src/Stepper.js
@@ -8,28 +8,7 @@ import Typography from '@material-ui/core/Typography';
 import SelectLocation from './assets/components/SelectLocation'
 import SelectFrontliner from './assets/components/SelectFrontliner'
 import WriteNote from './assets/components/WriteNote.js'
-import firebase from 'firebase';
 import { Box, Grid } from '@material-ui/core';
-
-//init firebase
-require('dotenv').config()
-console.log(process.env)
-
-const firebaseConfig = {
-  apiKey: process.env.REACT_APP_API,
-  authDomain: process.env.REACT_APP_DOM,
-  databaseURL: process.env.REACT_APP_DATABASE_URL,
-  projectId:  process.env.REACT_APP_ID,
-  storageBucket:  process.env.REACT_APP_BUCKET,
-  messagingSenderId:  process.env.REACT_APP_SENDER,
-  appId:  process.env.REACT_APP_APP_ID,
-  measurementId:  process.env.REACT_APP_MEASUREMENT
-};
-
-firebase.initializeApp(firebaseConfig);
-
-//init collection
-var formRef = firebase.database().ref('formData')
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -83,7 +62,7 @@ function getSteps() {
 }
 
 
-export default function ProgressStepper() {
+export default function ProgressStepper(props) {
   const classes = useStyles();
   const [activeStep, setActiveStep] = React.useState(0);
   const [completed, setCompleted] = React.useState({});
@@ -98,6 +77,8 @@ export default function ProgressStepper() {
   const [senderName, setSenderName] = React.useState(null);
 
   const steps = getSteps();
+
+  let formRef = props.firebase.db.ref('formData');
 
 React.useEffect(()=>{
     if(noteContent !== null && senderName !==null && noteContent !== "" && senderName !== ""){
@@ -244,7 +225,7 @@ React.useEffect(()=>{
           </Step>
         ))}
       </Stepper>
-      <div class={classes.cards}>
+      <div className={classes.cards}>
 
         {allStepsCompleted() ? (
           <Grid container
