+++ src/assets/components/FireBase.js
@@ -1,0 +1,6 @@
+import FireBase from './FireBase/FireBase';
+import FireBaseContext from './FireBase/FireBaseContext';
+
+export default FireBase;
+
+export { FireBaseContext };
