+++ src/assets/components/FireBase/FireBaseContext.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const FirebaseContext = React.createContext(null);
+
+export default FirebaseContext;
